Removed SubsetMethods
This module has always been a little half-baked.  There are other,
more standard ways of running a subset of tests (by directory,
using the NAME option with the runners, etc) and more flexible
ways of filtering a series of tests (ie using standard methdods).
It was syntax for the sake of syntax - so no more.