fix timeout into a variable
This appears to help prevent sporadic timeout errors, at least on
OSX.  Underlying is likely an issue within PTY spawn (it appears
bin/sh itself is not launching correctly because the first PS1 is
absent) or perhaps in a scheduler... in the kernel?  No concete
reason why this works.  Hopefully more later.