expect ps1 rather than nl to clear stty
It appears this is the cause of the sporadic missing ps1 prompts.
By logging the expects from this section it appears that in some
cases the ps1 could be eaten up by agent.expect("\n").  It is
more reliable to wait for the new prompt then clear it with a nl.